Fixed last.date EOF exception in process_data.py and update_db.py
Sometimes there was an EOF exception while parsing the last.dates file, which would crash the program, probably because another script was writing it at the same time. Now the program catches the error, prints a message and waits for another check; it will exit cleanly after the number of failed attempts specified in the setup files.